+++ transmitter/transmitter.ino
@@ -27,29 +27,27 @@ void loop()
 { 
   unsigned joystickXValue = analogRead(joystickXPin);
   unsigned joystickYValue = analogRead(joystickYPin);
-  int8_t value;
+  int8_t value[2];
 
   if (joystickYValue >= middleJoystickYValue)
   {
-    value = map(joystickYValue, middleJoystickYValue, 1023, 0, 127);
+    value[0] = map(joystickYValue, middleJoystickYValue, 1023, 0, 127);
   }
   else
   {
-    value = map(joystickYValue, 0, middleJoystickYValue, -128, 0);
+    value[0] = map(joystickYValue, 0, middleJoystickYValue, -128, 0);
   }
-
-  radio.transmit((uint8_t)value);
 
   if (joystickXValue >= middleJoystickXValue)
   {
-    value = map(joystickXValue, middleJoystickXValue, 1023, 0, 127);
+    value[1] = map(joystickXValue, middleJoystickXValue, 1023, 0, 127);
   }
   else
   {
-    value = map(joystickXValue, 0, middleJoystickXValue, -128, 0);
+    value[1] = map(joystickXValue, 0, middleJoystickXValue, -128, 0);
   }
   
-  radio.transmit((uint8_t)value);
+  radio.transmit((uint8_t*)value,2);
 
   delay(100);
 }
